Refactor: Restructure project with src layout and update README
