+++ .gitignore
@@ -2,3 +2,5 @@ build/
 dist/
 __pycache__/
 StudyWith.spec
+core
+core.*
