+++ extensions/chrome/background.js
