+++ extensions/chrome/manifest.json
