+++ requirements.txt
@@ -1,0 +1,4 @@
+Flask
+flask-cors
+psutil
+PyQt6
