+++ src/study_with/__init__.py
@@ -1,0 +1,5 @@
+"""Study With - Focus Timer package."""
+
+__all__ = ["__version__"]
+
+__version__ = "0.1.0"
