+++ src/study_with/__main__.py
@@ -1,0 +1,5 @@
+from .app import main
+
+
+if __name__ == "__main__":
+    main()
